+++ src/DashboardUI/src/pages/application/create/ApplicationCreatePage.tsx
@@ -4,7 +4,7 @@ import { ApplicationNavbar } from '../components/ApplicationNavbar';
 import ApplicationSubmissionForm from '../components/ApplicationSubmissionForm';
 import { ApplicationReviewPerspective } from '../../../data-contracts/ApplicationReviewPerspective';
 import ApplicationReviewHeader from '../components/ApplicationReviewHeader';
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 import ApplicationDocumentUploadSection from '../components/ApplicationDocumentUploadSection';
 import Button from 'react-bootstrap/esm/Button';
 
@@ -14,13 +14,28 @@ export function ApplicationCreatePage() {
   const { state } = useConservationApplicationContext();
   const navigate = useNavigate();
 
+  const formRef = useRef<HTMLFormElement>(null);
+  const [formValidated, setFormValidated] = useState(false);
   const [documentUploading, setDocumentUploading] = useState(false);
 
   const navigateToEstimationToolPage = () => {
     navigate(`/application/${state.conservationApplication.waterRightNativeId}/estimation`);
   };
 
-  const alertNotImplemented = () => alert('Not implemented. This feature will be implemented in a future release.');
+  const navigateToSubmitApplicationPage = () => {
+    navigate(`/application/${state.conservationApplication.waterConservationApplicationId}/submit`);
+  };
+
+  const handleSubmitClicked = () => {
+    const form = formRef.current;
+    const isFormValid = form!.checkValidity();
+
+    setFormValidated(true);
+
+    if (isFormValid) {
+      navigateToSubmitApplicationPage();
+    }
+  };
 
   return (
     <div className="d-flex flex-column flex-grow-1 h-100">
@@ -33,12 +48,12 @@ export function ApplicationCreatePage() {
       <div className="overflow-y-auto">
         <main className="container">
           <ApplicationReviewHeader perspective={perspective} />
-          <ApplicationSubmissionForm perspective={perspective} documentUploading={documentUploading} />
+          <ApplicationSubmissionForm ref={formRef} formValidated={formValidated} />
           <ApplicationDocumentUploadSection
             perspective={perspective}
             documentUploadProps={{ documentUploadingHandler: setDocumentUploading }}
           />
-          <ApplicantButtonRow documentUploading={documentUploading} handleSubmitClicked={alertNotImplemented} />
+          <ApplicantButtonRow documentUploading={documentUploading} handleSubmitClicked={handleSubmitClicked} />
         </main>
       </div>
     </div>
